# backend/docs/swagger.yaml
basePath: /admin-api
definitions:
  github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      msg:
        example: ""
        type: string
    type: object
...
  internal_modules_system.TenantSaveRequest:
    properties:
      accountCount:
        type: integer
      contactMobile:
        type: string
      contactName:
        type: string
      domain:
        type: string
      expireTime:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      status:
        type: integer
      username:
        type: string
      websites:
        items:
          type: string
        type: array
    required:
    - name
    type: object
...
info:
  contact: {}
  description: 'Go modular-monolith scaffold for operations platforms. Default database:
    MySQL 8.4.'
  title: Nimbus Framework Go API
  version: "1.0"
paths:
...
  /system/tenant/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_system.TenantSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Create a tenant and its initial administrator
      tags:
      - System Tenant
  /system/tenant/delete:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Delete a tenant and tenant-owned system data
      tags:
      - System Tenant
  /system/tenant/delete-list:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Delete tenants in batch
      tags:
      - System Tenant
...
  /system/tenant/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Get a tenant
      tags:
      - System Tenant
...
  /system/tenant/page:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Page tenants
      tags:
      - System Tenant
  /system/tenant/simple-list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List enabled tenants
      tags:
      - System Tenant
  /system/tenant/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Tenant
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_system.TenantSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Update a tenant
      tags:
      - System Tenant
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"